add `plusminus` to level4 of M per XKB update
The `plusminus` keysym is the last Latin-1 mathematical operator left
unbound on the standard us-intl layout, and likewise, M is the last
key position whose level4 shift state is the same as its level3 shift
state. This patch binds `plusminus` to the level4 shift state of M,
to go with a plausible phonetic nomenclature of plus«M»inus going on
M for Minus, just as how «m»icro is bound.